+++ ft_strjoin.c
@@ -13,8 +13,8 @@
 
 char	*ft_strjoin(char const *s1, char const *s2)
 {
-	int		len;
-	char	*dest;
+	size_t		len;
+	char		*dest;
 
 	if (!s1 || !s2)
 		return (0);
